+++ activestorage/test/controllers/direct_uploads_controller_test.rb
@@ -22,7 +22,10 @@ if SERVICE_CONFIGURATIONS[:s3] && SERVICE_CONFIGURATIONS[:s3][:access_key_id].pr
         "my_key_1": "my_value_1",
         "my_key_2": "my_value_2",
         "platform": "my_platform",
-        "library_ID": "12345"
+        "library_ID": "12345",
+        custom: {
+          "my_key_3": "my_value_3"
+        }
       }
 
       ActiveStorage::DirectUploadToken.stub(:verify_direct_upload_token, "s3") do
@@ -39,7 +42,7 @@ if SERVICE_CONFIGURATIONS[:s3] && SERVICE_CONFIGURATIONS[:s3][:access_key_id].pr
         assert_equal "text/plain", details["content_type"]
         assert_match SERVICE_CONFIGURATIONS[:s3][:bucket], details["direct_upload"]["url"]
         assert_match(/s3(-[-a-z0-9]+)?\.(\S+)?amazonaws\.com/, details["direct_upload"]["url"])
-        assert_equal({ "Content-Type" => "text/plain", "Content-MD5" => checksum, "Content-Disposition" => "inline; filename=\"hello.txt\"; filename*=UTF-8''hello.txt" }, details["direct_upload"]["headers"])
+        assert_equal({ "Content-Type" => "text/plain", "Content-MD5" => checksum, "Content-Disposition" => "inline; filename=\"hello.txt\"; filename*=UTF-8''hello.txt", "x-amz-meta-my_key_3" => "my_value_3" }, details["direct_upload"]["headers"])
       end
     end
   end
@@ -67,7 +70,10 @@ if SERVICE_CONFIGURATIONS[:gcs]
         "my_key_1": "my_value_1",
         "my_key_2": "my_value_2",
         "platform": "my_platform",
-        "library_ID": "12345"
+        "library_ID": "12345",
+        custom: {
+          "my_key_3": "my_value_3"
+        }
       }
 
       ActiveStorage::DirectUploadToken.stub(:verify_direct_upload_token, "gcs") do
@@ -83,7 +89,7 @@ if SERVICE_CONFIGURATIONS[:gcs]
         assert_equal metadata, details["metadata"].transform_keys(&:to_sym)
         assert_equal "text/plain", details["content_type"]
         assert_match %r{storage\.googleapis\.com/#{@config[:bucket]}}, details["direct_upload"]["url"]
-        assert_equal({ "Content-MD5" => checksum, "Content-Disposition" => "inline; filename=\"hello.txt\"; filename*=UTF-8''hello.txt" }, details["direct_upload"]["headers"])
+        assert_equal({ "Content-MD5" => checksum, "Content-Disposition" => "inline; filename=\"hello.txt\"; filename*=UTF-8''hello.txt", "x-goog-meta-my_key_3" => "my_value_3" }, details["direct_upload"]["headers"])
       end
     end
   end